Added VectorLayer to the Map
