+++ src/components/Map/Source/vector.js
@@ -1,0 +1,9 @@
+import { Vector as VectorSource } from 'ol/source';
+
+function vector({ features }) {
+	return new VectorSource({
+		features
+	});
+}
+
+export default vector;
